Extract tracker label from tracker card
 - this should allow the tracker pattern to be used without forcing the
   use of TrackerCard component
