+++ packages/libs/react-ui/src/components/TrackerCard/index.ts
@@ -1,1 +1,5 @@
 export { TrackerCard } from './TrackerCard';
+export { TrackerLabel } from './TrackerLabel';
+
+export type { ITrackerCardProps } from './TrackerCard';
+export type { ILabelValue } from './TrackerLabel';
