+++ packages/libs/react-ui/src/index.ts
@@ -116,6 +116,7 @@ export {
   TextareaField,
   Tooltip,
   TrackerCard,
+  TrackerLabel,
   Tree,
   maskValue,
   useDialog,
